requirements.txt
zeroc-ice
xarray
pandas
urllib3
numpy